pyproject.toml
[project]
name = "univocalx-dj"
version = "0.2.7a0"
description = ""
authors = [
    {name = "elad jerbi",email = "eladxjerbi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic (>=2.11.7,<2.12.0)",
    "pydantic-settings (>=2.10.1,<2.11.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "boto3 (>=1.39.16,<2.0.0)",
    "sqlalchemy (>=2.0.42,<2.1.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
]

include = [
    { path = "assets/*", format = ["sdist", "wheel"] }
]

[tool.poetry.group.dev.dependencies]
ruff = "=0.12.5"
mypy = "=1.9.0"
pytest = "=8.0.0"
pytest-cov = "=4.1.0"

[tool.poetry]
packages = [{include = "dj", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dj = "dj.main:main"


[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4

target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
extend-select = ["I"]

ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

indent-style = "space"

skip-magic-trailing-comma = false

line-ending = "auto"

docstring-code-format = false

docstring-code-line-length = "dynamic"

# Mypy configuration
[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
check_untyped_defs = true
disable_error_code = ["import-untyped"]